pyproject.toml
[project]
name = "algobattle_web"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Imogen Hergeth"}, {name = "Henri Lotze"}]
urls = {github = "https://github.com/Benezivas/algobattle-web"}
classifiers = [
    "Natural Language :: English",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Typing :: Typed",
]
dependencies = [
    "algobattle",
    "fastapi",
    "jinja2",
    "python-jose[cryptography]",
    "sqlalchemy",
    "sqlalchemy_utils",
]

[tool.pyright]
diagnosticMode = "workspace"
